requirements.txt
ipykernel==6.22.0
jupyter==1.0.0
openai==0.27.6
panel==0.14.4
python-dotenv==1.0.0
redlines==0.2.2